admin sepeku prevented

+++ src/main/java/main/WTLibraryApp/User/UserController.java
@@ -118,9 +118,15 @@ public class UserController {
 
 //	Deletes an user from the table.
 	@GetMapping("/users/delete-user/{id}")
-	public String deleteUser(@PathVariable("id") long id, Model model) {
-		User users = service.findUser(id);
-		service.deleteUser(users, id);
-		return "redirect:/users";
+	public String deleteUser(@PathVariable("id") long id, @CurrentSecurityContext(expression = "authentication") Authentication authentication, Model model) {
+		User user = service.findUser(id);
+		User currentUser = service.findByEmail(authentication.getName());
+		if(user==currentUser) {
+			return "redirect:/users/edit-user/" + id + "?sepeku";
+		}else {
+			User users = service.findUser(id);
+			service.deleteUser(users, id);
+			return "redirect:/users";
+		}
 	}
 }
